Set required parameters for URLs in the secondary hero component
