+++ src/amo/components/SecondaryHero/index.js
@@ -2,18 +2,42 @@
 import invariant from 'invariant';
 import * as React from 'react';
 
-import type { SecondaryHeroShelfType } from 'amo/reducers/home';
+import { addParamsToHeroURL, isInternalURL } from 'amo/utils';
+import type {
+  HeroCallToActionType,
+  SecondaryHeroShelfType,
+} from 'amo/reducers/home';
 
 import './styles.scss';
 
+export const SECONDARY_HERO_SRC = 'homepage-secondary-hero';
+
 type Props = {| shelfData: SecondaryHeroShelfType |};
 
-export const SecondaryHeroBase = ({ shelfData }: Props) => {
+type InternalProps = {|
+  ...Props,
+  _isInternalURL: typeof isInternalURL,
+|};
+
+export const SecondaryHeroBase = ({
+  _isInternalURL = isInternalURL,
+  shelfData,
+}: InternalProps) => {
+  const makeCallToActionURL = (urlString: string) => {
+    return addParamsToHeroURL({ heroSrcCode: SECONDARY_HERO_SRC, urlString });
+  };
+
   const { headline, description, cta, modules } = shelfData;
 
   invariant(headline, 'The headline property is required');
   invariant(description, 'The description property is required');
   invariant(modules, 'The modules property is required');
+
+  const getLinkProps = (link: HeroCallToActionType | null) => {
+    return !link || _isInternalURL({ urlString: link.url })
+      ? {}
+      : { rel: 'noopener noreferrer', target: '_blank' };
+  };
 
   const renderedModules = [];
   modules.forEach((module) => {
@@ -28,7 +52,11 @@ export const SecondaryHeroBase = ({ shelfData }: Props) => {
           {module.description}
         </div>
         {module.cta && (
-          <a className="SecondaryHero-module-link" href={module.cta.url}>
+          <a
+            className="SecondaryHero-module-link"
+            href={makeCallToActionURL(module.cta.url)}
+            {...getLinkProps(module.cta)}
+          >
             <span className="SecondaryHero-module-linkText">
               {module.cta.text}
             </span>
@@ -44,7 +72,11 @@ export const SecondaryHeroBase = ({ shelfData }: Props) => {
         <h2 className="SecondaryHero-message-headline">{headline}</h2>
         <div className="SecondaryHero-message-description">{description}</div>
         {cta && (
-          <a className="SecondaryHero-message-link" href={cta.url}>
+          <a
+            className="SecondaryHero-message-link"
+            href={makeCallToActionURL(cta.url)}
+            {...getLinkProps(cta)}
+          >
             <span className="SecondaryHero-message-linkText">{cta.text}</span>
           </a>
         )}
