+++ tests/unit/amo/components/TestHeroRecommendation.js
@@ -1,15 +1,12 @@
-import url from 'url';
-
 import * as React from 'react';
 
 import HeroRecommendation, {
   PRIMARY_HERO_CLICK_CATEGORY,
   PRIMARY_HERO_SRC,
-  addParamsToHeroURL,
   HeroRecommendationBase,
 } from 'amo/components/HeroRecommendation';
 import { createInternalHeroShelves } from 'amo/reducers/home';
-import { getAddonURL } from 'amo/utils';
+import { addParamsToHeroURL, getAddonURL } from 'amo/utils';
 import {
   createFakeEvent,
   createFakeTracking,
@@ -17,7 +14,6 @@ import {
   fakeAddon,
   fakeI18n,
   fakePrimaryHeroShelfExternal,
-  getFakeConfig,
   shallowUntilTarget,
 } from 'tests/unit/helpers';
 
@@ -166,114 +162,6 @@ describe(__filename, () => {
     expect(root.find('.HeroRecommendation-body').html()).toContain(description);
   });
 
-  describe('addParamsToHeroURL', () => {
-    let _addQueryParams;
-    let _isInternalURL;
-    const urlString = '/path/name';
-    const internalQueryParams = { internalParam1: 'internalParam1' };
-    const externalQueryParams = { externalParam1: 'externalParam1' };
-
-    beforeEach(() => {
-      _addQueryParams = sinon.spy();
-      _isInternalURL = sinon.stub();
-    });
-
-    it('passes internal query params to _addQueryParams for an internal URL', () => {
-      _isInternalURL.returns(true);
-
-      addParamsToHeroURL({
-        _addQueryParams,
-        _isInternalURL,
-        externalQueryParams,
-        internalQueryParams,
-        urlString,
-      });
-
-      sinon.assert.calledWith(_addQueryParams, urlString, internalQueryParams);
-    });
-
-    it('passes default internal query params to _addQueryParams for an internal URL', () => {
-      _isInternalURL.returns(true);
-
-      addParamsToHeroURL({
-        _addQueryParams,
-        _isInternalURL,
-        urlString,
-      });
-
-      sinon.assert.calledWith(_addQueryParams, urlString, {
-        src: PRIMARY_HERO_SRC,
-      });
-    });
-
-    it('allows for override of heroSrcCode for an internal URL', () => {
-      const heroSrcCode = 'test-src-code';
-      _isInternalURL.returns(true);
-
-      addParamsToHeroURL({
-        _addQueryParams,
-        _isInternalURL,
-        heroSrcCode,
-        urlString,
-      });
-
-      sinon.assert.calledWith(_addQueryParams, urlString, {
-        src: heroSrcCode,
-      });
-    });
-
-    it('passes external query params to _addQueryParams for an external URL', () => {
-      _isInternalURL.returns(false);
-
-      addParamsToHeroURL({
-        _addQueryParams,
-        _isInternalURL,
-        externalQueryParams,
-        internalQueryParams,
-        urlString,
-      });
-
-      sinon.assert.calledWith(_addQueryParams, urlString, externalQueryParams);
-    });
-
-    it('passes default external query params to _addQueryParams for an external URL', () => {
-      const baseURL = 'https://example.org';
-      const _config = getFakeConfig({ baseURL });
-      _isInternalURL.returns(false);
-
-      addParamsToHeroURL({
-        _addQueryParams,
-        _config,
-        _isInternalURL,
-        urlString,
-      });
-
-      sinon.assert.calledWith(_addQueryParams, urlString, {
-        utm_content: PRIMARY_HERO_SRC,
-        utm_medium: 'referral',
-        utm_source: url.parse(baseURL).host,
-      });
-    });
-
-    it('allows for override of heroSrcCode for an external URL', () => {
-      const heroSrcCode = 'test-src-code';
-      _isInternalURL.returns(false);
-
-      addParamsToHeroURL({
-        _addQueryParams,
-        _isInternalURL,
-        heroSrcCode,
-        urlString,
-      });
-
-      sinon.assert.calledWith(
-        _addQueryParams,
-        urlString,
-        sinon.match({ utm_content: heroSrcCode }),
-      );
-    });
-  });
-
   describe('makeCallToActionURL', () => {
     it('creates a URL for an addon', () => {
       const slug = 'some-addon-slug';
@@ -282,7 +170,10 @@ describe(__filename, () => {
       const root = render({ shelfData });
 
       expect(root.instance().makeCallToActionURL()).toEqual(
-        addParamsToHeroURL({ urlString: getAddonURL(slug) }),
+        addParamsToHeroURL({
+          heroSrcCode: PRIMARY_HERO_SRC,
+          urlString: getAddonURL(slug),
+        }),
       );
     });
 
@@ -295,7 +186,10 @@ describe(__filename, () => {
       const root = render({ shelfData });
 
       expect(root.instance().makeCallToActionURL()).toEqual(
-        addParamsToHeroURL({ urlString: homepage }),
+        addParamsToHeroURL({
+          heroSrcCode: PRIMARY_HERO_SRC,
+          urlString: homepage,
+        }),
       );
     });
   });
